+++ py_gql/utilities/__init__.py
@@ -19,6 +19,7 @@ from .collect_fields import (
     selected_fields,
 )
 from .introspection_query import introspection_query
+from .max_depth import MaxDepthValidationRule
 from .type_info import TypeInfoVisitor
 from .untyped_value_from_ast import untyped_value_from_ast
 from .value_from_ast import value_from_ast
@@ -37,4 +38,5 @@ __all__ = (
     "TypeInfoVisitor",
     "untyped_value_from_ast",
     "value_from_ast",
+    "MaxDepthValidationRule",
 )